Migrate some code from PerformanceController to PerformanceLive
